requirements.txt
# client
requests

# server
flask
psycopg2

# mqtt
paho-mqtt